requirements.txt
pyqt5~=5.15.6
pillow~=8.4.0